enhance(frontend): 設定の移行を手動でトリガーできるように
