+++ locales/index.d.ts
@@ -5514,6 +5514,14 @@ export interface Locale extends ILocale {
      * チャット
      */
     "chat": string;
+    /**
+     * 旧設定情報を移行
+     */
+    "migrateOldSettings": string;
+    /**
+     * 通常これは自動で行われていますが、何らかの理由により上手く移行されなかった場合は手動で移行処理をトリガーできます。現在の設定情報は上書きされます。
+     */
+    "migrateOldSettings_description": string;
     "_chat": {
         /**
          * まだメッセージはありません
